refactor(anvil): isolate Monad context helpers
